Polyline and vector
