+++ Rhinoceros_Engine/Create/PolylineCurve.cs
@@ -22,10 +22,6 @@
 
 using Rhino.Geometry;
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace BH.Engine.Rhinoceros
 {
@@ -35,30 +31,16 @@ namespace BH.Engine.Rhinoceros
         /**** Public Methods                            ****/
         /***************************************************/
 
-        public static Vector3d RandomVector3d(Random random)
+        public static PolylineCurve RandomPolylineCurve(Random random)
         {
-            return new Vector3d(random.NextDouble(), random.NextDouble(), random.NextDouble());
+            return new PolylineCurve(RandomPoint3dList(random));
         }
 
         /***************************************************/
 
-        public static Vector3f RandomVector3f(Random random)
+        public static PolylineCurve RandomPolylineCurve(int seed = 0)
         {
-            return new Vector3f((float)random.NextDouble(), (float)random.NextDouble(), (float)random.NextDouble());
-        }
-
-        /***************************************************/
-
-        public static Vector3d RandomVector3d(int seed = 0)
-        {
-            return RandomVector3d(new Random(seed));
-        }
-
-        /***************************************************/
-
-        public static Vector3f RandomVector3f(int seed = 0)
-        {
-            return RandomVector3f(new Random(seed));
+            return RandomPolylineCurve(new Random(seed));
         }
 
         /***************************************************/
