+++ Rhinoceros_Engine/Create/Vector3f.cs
@@ -31,16 +31,16 @@ namespace BH.Engine.Rhinoceros
         /**** Public Methods                            ****/
         /***************************************************/
 
-        public static Polyline RandomPolyline(Random random)
+        public static Vector3f RandomVector3f(Random random)
         {
-            return new Polyline(RandomPoint3dList(random));
+            return new Vector3f((float)random.NextDouble(), (float)random.NextDouble(), (float)random.NextDouble());
         }
 
         /***************************************************/
 
-        public static Polyline RandomPolyline(int seed = 0)
+        public static Vector3f RandomVector3f(int seed = 0)
         {
-            return RandomPolyline(new Random(seed));
+            return RandomVector3f(new Random(seed));
         }
 
         /***************************************************/
